refactor DownloadAsCSV logic out of line plot module
